RF: move "our" get_metadata to dandi.metadata, and safe_call into utils
