+++ dandi/cli/cmd_ls.py
@@ -6,6 +6,7 @@ from dandi.cli.command import get_files
 
 from .command import lgr, main
 
+from ..utils import safe_call
 
 # TODO: all the recursion options etc
 
@@ -118,20 +119,13 @@ def ls(paths, fields=None, format="auto"):
 
 def get_metadata_pyout(path, keys=None, process_paths=None):
     from ..pynwb_utils import (
-        get_metadata,
         get_nwb_version,
         get_neurodata_types,
         ignore_benign_pynwb_warnings,
     )
+    from ..metadata import get_metadata
 
     ignore_benign_pynwb_warnings()
-
-    def safe_call(func, path, default=None):
-        try:
-            return func(path)
-        except Exception as exc:
-            lgr.debug("Call to %s on %s failed: %s", func.__name__, path, exc)
-            return default
 
     def fn():
         rec = {}
