+++ dandi/metadata.py
@@ -1,0 +1,63 @@
+from dandi.pynwb_utils import metadata_cache, get_nwb_version, _get_pynwb_metadata
+
+from . import get_logger
+
+lgr = get_logger()
+
+
+@metadata_cache.memoize_path
+def get_metadata(path):
+    """Get selected metadata from a .nwb file
+
+    Parameters
+    ----------
+    path: str or Path
+
+    Returns
+    -------
+    dict
+    """
+    path = str(path)  # for Path
+    out = dict()
+
+    # First read out possibly available versions of specifications for NWB(:N)
+    out["nwb_version"] = get_nwb_version(path)
+
+    # PyNWB might fail to load because of missing extensions.
+    # There is a new initiative of establishing registry of such extensions.
+    # Not yet sure if PyNWB is going to provide "native" support for needed
+    # functionality: https://github.com/NeurodataWithoutBorders/pynwb/issues/1143
+    # So meanwhile, hard-coded workaround for data types we care about
+    ndtypes_registry = {
+        "AIBS_ecephys": "allensdk.brain_observatory.ecephys.nwb",
+        "ndx-labmetadata-abf": "ndx_dandi_icephys",
+    }
+    tried_imports = set()
+    while True:
+        try:
+            out.update(_get_pynwb_metadata(path))
+            break
+        except KeyError as exc:  # ATM there is
+            lgr.debug("Failed to read %s: %s", path, exc)
+            import re
+
+            res = re.match(r"^['\"\\]+(\S+). not a namespace", str(exc))
+            if not res:
+                raise
+            ndtype = res.groups()[0]
+            if ndtype not in ndtypes_registry:
+                raise ValueError(
+                    "We do not know which extension provides %s. "
+                    "Original exception was: %s. " % (ndtype, exc)
+                )
+            import_mod = ndtypes_registry[ndtype]
+            lgr.debug("Importing %r which should provide %r", import_mod, ndtype)
+            if import_mod in tried_imports:
+                raise RuntimeError(
+                    "We already tried importing %s to provide %s, but it seems it didn't help"
+                    % (import_mod, ndtype)
+                )
+            tried_imports.add(import_mod)
+            __import__(import_mod)
+
+    return out
